Version 0.0.2 * fix build * update SDK
