+++ .gitignore
@@ -1,4 +1,3 @@
-/build/qiwi.ocmod.zip
+/build
 /vendor
-/src/upload/system/library/qiwi/vendor
 composer.lock
