--- src/upload/catalog/language/ru-ru/extension/payment/qiwi.php
@@ -1,16 +1,0 @@
-<?php
-
-// Text
-$_['text_title']     = 'QIWI Касса';
-$_['button_confirm'] = 'Подтвердить заказ';
-$_['text_loading']   = 'Загрузка...';
-$_['create_bill']    = '<a href="%1$s" target="_blank">Счет</a> готов к оплате';
-$_['waiting_bill']   = 'Счет ожидает оплаты';
-$_['paid_bill']      = 'Счет оплаче';
-$_['rejected_bill']  = 'Счет отменен';
-$_['expired_bill']   = 'Счет истек';
-$_['partial_bill']   = 'Счет частично возвращен';
-$_['full_bill']      = 'Счет полностью возвращен';
-
-// Error
-$_['error_api']      = 'ОШИБКА: Платежный сервис QIWI Касса временно не доступен! Пожалуйста, попробуйте позже.';
