--- src/upload/system/library/qiwi/composer.json
@@ -1,16 +1,0 @@
-{
-  "require": {
-    "php": ">=5.6.0",
-    "qiwi/bill-payments-php-sdk": "^0"
-  },
-  "autoload": {
-    "psr-4": {
-      "Qiwi\\": "src/"
-    }
-  },
-  "config": {
-    "platform": {
-      "php": "5.6.0"
-    }
-  }
-}
