+++ upload/admin/language/en-gb/extension/payment/qiwi.php
@@ -1,0 +1,85 @@
+<?php
+
+/** @noinspection HtmlUnknownTarget */
+
+// Heading
+$_['heading_title']         = 'QIWI Kassa';
+
+// Text
+$_['text_qiwi']             = '<a target="_blank" href="https://kassa.qiwi.com/"><img src="view/image/payment/qiwi.png" alt="QIWI Kassa" title="QIWI Kassa" style="border: 1px solid #eee;"/></a>';
+$_['text_supports']         = 'Payment over: VISA, MasterCard, MIR, Phone balance, QIWI Wallet';
+$_['text_description']      = '<p>To start working with the QIWI cash service, you need to <a href="https://kassa.qiwi.com/" target="_blank">register a store</a>.</p>'
+                            . '<p>Also, a <a href="https://kassa.qiwi.com/" target="_blank">demonstration stand</a> is available for you.</p>';
+$_['text_extension']        = 'Extensions';
+$_['text_success']          = 'Success: You have modified QIWI Kassa payment module!';
+$_['text_edit']             = 'Edit QIWI Kassa';
+$_['text_enabled']          = 'Enabled';
+$_['text_disabled']         = 'Disabled';
+$_['text_popup_enabled']    = 'On the store page, via popup';
+$_['text_popup_disabled']   = 'On an external site, by redirect';
+$_['text_payment_info']     = 'Payment information';
+$_['text_from']             = 'From:';
+$_['text_to']               = 'To:';
+$_['text_bill_title']       = 'Bill';
+$_['text_bill_amount']      = 'Amount';
+$_['text_bill_status']      = 'Status';
+$_['text_bill_date']        = 'Alive';
+$_['text_bill_action']      = 'Actions';
+$_['text_reject']           = 'Reject';
+$_['text_reject_success']   = 'Invoice will rejecting.';
+$_['text_refund']           = 'Refund';
+$_['text_refund_success']   = 'Invoice will refunding.';
+
+// Tabs
+$_['tab_general']           = 'General settings';
+$_['tab_qiwi']              = 'Qiwi Kassa integration';
+$_['tab_order_status']      = 'Order status';
+
+// Entry
+$_['entry_status']          = 'Status';
+$_['entry_title']           = 'Title';
+$_['entry_description']     = 'Description';
+$_['entry_sort_order']      = 'Sort order';
+$_['entry_total']           = 'Total';
+$_['entry_geo_zone']        = 'Geo zone';
+$_['entry_notification']    = 'Notification address';
+$_['entry_key_secret']      = 'Secret key';
+$_['entry_key_public']      = 'Public key';
+$_['entry_theme_code']      = 'Theme style code';
+$_['entry_popup']           = 'Payment form';
+$_['entry_live_time']       = 'Live time';
+$_['entry_debug']           = 'Debug mode';
+$_['entry_waiting_status']  = 'Payment waiting status';
+$_['entry_paid_status']     = 'Paid status';
+$_['entry_rejected_status'] = 'Rejected status';
+$_['entry_expired_status']  = 'Expired status';
+$_['entry_partial_status']  = 'Partial refunded status';
+$_['entry_full_status']     = 'Full refunded status';
+$_['entry_refund_amount']   = 'Refund amount';
+
+// Help
+$_['help_status']           = '';
+$_['help_title']            = 'The name of the payment method displayed to customers.';
+$_['help_description']      = 'Description of the payment method displayed to customers.';
+$_['help_sort_order']       = '';
+$_['help_total']            = 'The checkout total the order must reach before this payment method becomes active.';
+$_['help_geo_zone']         = '';
+$_['help_notification']     = 'Set this value in the payment system store settings.';
+$_['help_key_secret']       = 'The secret key to the payment system for your store.';
+$_['help_key_public']       = 'The public key to the payment system for your store.';
+$_['help_theme_code']       = 'Personalization code of the payment form style is presented in the payment system store settings.';
+$_['help_live_time']        = 'The live time for unpaid invoice in days.';
+$_['help_popup']            = '';
+$_['help_debug']            = 'Log api requests.';
+$_['help_waiting_status']   = '';
+$_['help_paid_status']      = '';
+$_['help_rejected_status']  = '';
+$_['help_expired_status']   = '';
+$_['help_partial_status']   = '';
+$_['help_full_status']      = '';
+
+// Error
+$_['error_permission']       = 'Warning: You do not have permission to modify payment QIWI Kassa!';
+$_['error_order_permission'] = 'Warning: You do not have permission to modify order!';
+$_['error_required']         = 'Warning: Сhanges not saved because required settings not set!';
+$_['error_required_field']   = 'This field required.';
