+++ upload/catalog/language/en-gb/extension/payment/qiwi.php
@@ -1,0 +1,16 @@
+<?php
+
+// Text
+$_['text_title']     = 'QIWI Kassa';
+$_['button_confirm'] = 'Confirm Order';
+$_['text_loading']   = 'Loading...';
+$_['create_bill']    = '<a href="%1$s" target="_blank">Invoice</a> ready for payment';
+$_['waiting_bill']   = 'Invoice await payment';
+$_['paid_bill']      = 'Invoice paid';
+$_['rejected_bill']  = 'Invoice rejected';
+$_['expired_bill']   = 'Invoice expired';
+$_['partial_bill']   = 'Invoice partial refunded';
+$_['full_bill']      = 'Invoice full refunded';
+
+// Error
+$_['error_api']      = 'ERROR: Payment service QIWI Kassa temporary not available! Please try later.';
